+++ redwood-tooling-codegen/src/main/kotlin/app/cash/redwood/tooling/codegen/modifierGeneration.kt
@@ -17,6 +17,7 @@ package app.cash.redwood.tooling.codegen
 
 import app.cash.redwood.tooling.schema.Modifier
 import app.cash.redwood.tooling.schema.Schema
+import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.TypeSpec
@@ -27,7 +28,20 @@ internal fun generateModifierInterface(schema: Schema, modifier: Modifier): File
     .addAnnotation(suppressDeprecations)
     .addType(
       TypeSpec.interfaceBuilder(type)
-        .addSuperinterface(Redwood.ModifierElement)
+        .addSuperinterface(
+          if (modifier.scopes.isEmpty()) {
+            Redwood.ModifierUnscopedElement
+          } else {
+            Redwood.ModifierScopedElement
+          },
+        )
+        .optIn(Stdlib.ExperimentalObjCName)
+        .addAnnotation(
+          AnnotationSpec.builder(Stdlib.ObjCName)
+            .addMember("%S", type.simpleName)
+            .addMember("exact = true")
+            .build(),
+        )
         .maybeAddDeprecation(modifier.deprecation)
         .apply {
           for (property in modifier.properties) {