+++ test-app/ios-uikit/TestApp/IosTestSchemaWidgetFactory.swift
@@ -47,4 +47,19 @@ class IosTestSchemaWidgetFactory: TestSchemaWidgetFactory {
     func TestRow() -> TestRow {
         fatalError()
     }
+
+    func BackgroundColor(value: Any, modifier: BackgroundColor) {
+        (value as! UIView).backgroundColor = UIColor(argb: UInt(modifier.color))
+    }
+}
+
+extension UIColor {
+    convenience init(argb: UInt) {
+        let alpha = CGFloat((argb >> 24) & 0xFF) / 255.0
+        let red = CGFloat((argb >> 16) & 0xFF) / 255.0
+        let green = CGFloat((argb >> 8) & 0xFF) / 255.0
+        let blue = CGFloat(argb & 0xFF) / 255.0
+
+        self.init(red: red, green: green, blue: blue, alpha: alpha)
+    }
 }