+++ test-app/ios-uikit/TestApp/RectangleBinding.swift
@@ -36,14 +36,3 @@ class RectangleBinding: Rectangle {
         root.layer.cornerRadius = CGFloat(cornerRadius)
     }
 }
-
-private extension UIColor {
-    convenience init(argb: UInt) {
-        let alpha = CGFloat((argb >> 24) & 0xFF) / 255.0
-        let red = CGFloat((argb >> 16) & 0xFF) / 255.0
-        let green = CGFloat((argb >> 8) & 0xFF) / 255.0
-        let blue = CGFloat(argb & 0xFF) / 255.0
-
-        self.init(red: red, green: green, blue: blue, alpha: alpha)
-    }
-}