+++ test-app/presenter/src/commonMain/kotlin/com/example/redwood/testing/presenter/TestApp.kt
@@ -36,6 +36,7 @@ private val screens = buildMap<String, @Composable TestContext.() -> Unit> {
   put("Repo Search") { RepoSearch(httpClient) }
   put("UI Configuration") { UiConfigurationValues() }
   put("Box Sandbox") { BoxSandbox() }
+  put("Unscoped Modifiers") { UnscopedModifiers() }
 }
 
 @Stable